Basic JS Lessons Day 1 - Next day push
I finished these yesterday and now I’m pushing them. All are correct as
of the date they are published, but always remember FCC updates their
exercises regularly so you need to be ready with your own solution and
only use these as a guide. I know I made an error on an assumption I
made with the counting cards js exercise and that can happen to anyone.